mage_ai/blackbirdnest/requirements.txt
selenium==4.11
requests
python-slugify
tenacity==8.2.2
unicode
pytest-playwright